Add FM analog RX "stuff".
fm_mode(freq) will enter analog FM mode, optionally setting a
frequency.  Returns false if the frequency can't be set.

set_freq(freq, high_res) sets the RX frequency to freq MHz, optionally
using the higher power but more precise 18-bit divisor mode (doesn't
look like it's necessary really).  Returns true on success.  This
uses the frequency offset my radio happens to show.  It is *not*
centred on XF501 which suggests I need to parse the calibration data
to get the correct values.

get_freq() returns the last successfully set frequency in MHz.  It's
possible the frequency is lost when init_pins() is called since this
likely powers down the frequency synthesizer.  This needs to be
investigated and cleaned up if necessary.

scan(start, end, step) scans frequencies from start MHz to end MHz
using the specified step.  It returns when the signal strength is
greater than 0.4.  Again, this needs to parse calibration data and
likely the squelch setting.  Also, sitting in a loop like that isn't
really a good thing to do, but meh.

None of these functions range check parameters, so you can ask for
the impossible.  This also needs to be fixed up.