Slow the LCD down more...
It seems that the issue with using FSMC is that we can't control
the data hold time.  To work aorund this, switch to bit-banging it
instead of slowing down the system clock.

There's a few optimizations in here that tried to get FSMC working,
but they mostly don't matter.

As usual, readPixel wasn't tested.

If this is simply too damn slow, there's some other optimizations
that can be done, but this should be good enough for now.